Moved rendering into a separate class
