+++ lib/govuk_tech_docs/api_reference/templates/api_reference_full.html.erb
@@ -1,5 +1,5 @@
 <h1 id="<%=  info.title.parameterize %>"><%= info.title %> v<%= info.version %></h1>
-<%= render_markdown(info.description) %>
+<%= markdown(info.description) %>
 <% if server %>
 <h2 id="base-url">Base URL</h2>
 <p><strong><%= server.url %></strong></p>
