+++ lib/govuk_tech_docs/api_reference/templates/schema.html.erb
@@ -1,5 +1,5 @@
 <h3 id="<%= id = 'schema-' + title; id.parameterize %>"><%= title %></h3>
-<%= render_markdown(schema.description) %>
+<%= markdown(schema.description) %>
 <% if schema.properties.any? %>
 <table>
 <thead>
@@ -11,7 +11,7 @@
 <td><%= property[0] %></td>
 <td><%= property[1].type %></td>
 <td><%= property[1].required.present? %></td>
-<td><%= render_markdown(property[1].description) %></td>
+<td><%= markdown(property[1].description) %></td>
 </tr>
 <% end %>
 </tbody>
